# web/swagger.yaml
swagger: '2.0'
info:
  title: Mindwell
  version: "1.0"
basePath: /api/v1
produces:
  - application/json
securityDefinitions:
  OAuth2App:
    type: oauth2
    flow: application
    tokenUrl: "/oauth2/token"
  OAuth2Password:
    type: oauth2
    flow: password
    tokenUrl: "/oauth2/token"
  OAuth2Code:
    type: oauth2
    flow: accessCode
    authorizationUrl: "/oauth"
    tokenUrl: "/oauth2/token"
    scopes:
      "account:read": read available invites, etc.
      "account:write": change email, password, etc.
      "adm:read": read adm info and statuses
      "adm:write": change adm info and update statuses
      "comments:read": read comments and comment votes
      "comments:write": create, edit and remove comments
      "entries:read": read entries, entry votes, favorites, watchings, tags
      "entries:write": create, edit and remove tlog entries
      "favorites:write": add entries to favorites and remove
      "images:read": get uploaded images
      "images:write": upload new images
      "messages:read": read chats and messages and set them read
      "messages:write": create, edit and remove messages
      "notifications:read": read notifications and set them read
      "relations:write": follow, block, hide tlogs, accept followers, invite people
      "settings:read": read account settings
      "settings:write": change account settings
      "themes:read": read theme info, followers
      "themes:write": create, edit themes
      "users:read": read profile info, relations
      "users:write": change profile info, avatar, etc.
      "votes:write": vote and unvote entries and comments
      "watchings:write": subscribe to entry comments and unsubscribe
      "wishes:read": read received wishes
      "wishes:write": send new wishes
definitions:
...
  Error:
    type: object
    properties:
      message:
        type: string
    example:
      message: Access denied
...
  Color:
    type: string
    pattern: '^#[0-9a-fA-F]{6}$'
    example: '#373737'
    description: color in rgb
  Avatar:
    type: object
    properties:
      x124:
        type: string
      x92:
        type: string
      x42:
        type: string
  Cover:
    type: object
    properties:
      id:
        type: integer
        format: int64
        minimum: 1
        description: user id
      x1920:
        type: string
      x318:
        type: string
  User:
    type: object
    properties:
      id:
        type: integer
        format: int64
        minimum: 1
      name:
        type: string
        minLength: 1
        maxLength: 20
      showName:
        type: string
        minLength: 1
        maxLength: 20
      isTheme:
        type: boolean
      isOnline:
        type: boolean
      avatar:
        $ref: '#/definitions/Avatar'
    example:
      id: 1
      name: 'Mindwell'
      showName: 'Майндвелл'
      isOnline: false
  Friend: 
    allOf:
      - $ref: '#/definitions/User'
      - type: object
        properties:
          gender:
            type: string
            enum: [male, female, not set]
          privacy:
            type: string
            enum: [all, followers, invited, registered]
          chatPrivacy:
            type: string
            enum: [invited, followers, friends, me]
          title:
            type: string
            maxLength: 260
          rank:
            type: number
            format: int64
          lastSeenAt:
            type: number
            format: double
          cover:
            $ref: '#/definitions/Cover'
          counts:
            type: object
            properties:
              entries:
                type: integer
              followings:
                type: integer
              followers:
                type: integer
              ignored:
                type: integer
              invited:
                type: integer
              comments:
                type: integer
              favorites:
                type: integer
              tags:
                type: integer
              days:
                type: integer
...
  Profile:
    allOf:
      - $ref: '#/definitions/Friend'
      - type: object
        properties:
          isDaylog:
            type: boolean
          createdAt:
            type: number
            format: double
          invitedBy:
            $ref: '#/definitions/User'
          createdBy:
            $ref: '#/definitions/User'
          ageLowerBound:
            type: integer
          ageUpperBound:
            type: integer
          country:
            type: string
            maxLength: 50
          city:
            type: string
            maxLength: 50
          design:
            $ref: '#/definitions/Design'
          relations:
            type: object
            properties:
              isOpenForMe:
                type: boolean
              toMe:
                type: string
                enum: [followed, requested, ignored, hidden, none]
              fromMe:
                type: string
                enum: [followed, requested, ignored, hidden, none]
          rights:
            type: object
            properties:
              chat:
                type: boolean
              ignore:
                type: boolean
              complain:
                type: boolean
        example:
          gender: male
          isDaylog: false
          privacy: all
          title: 'some tlog title'
          rank: 17
          createdAt: 1531029717.333
          lastSeenAt: 1531029717.333
          country: Россия
          city: Казань
          counts:
            entries: 13
            followings: 4
            followers: 15
            ignored: 0
            invited: 0
            comments: 22
            favorites: 1
            tags: 0
          relations:
            toMe: none
            fromMe: followed
  Design:
    type: object
    properties:
      css:
        type: string
      backgroundColor:
        $ref: '#/definitions/Color'
      textColor:
        $ref: '#/definitions/Color'
      fontFamily:
        type: string
      fontSize:
        type: integer
      textAlignment:
        type: string
        enum: [left, right, center, justify]
...
parameters:
...
  pathName:
    in: path
    name: name
    type: string
    minLength: 1
    maxLength: 20
    required: true
    pattern: "^[0-9\\-_]*[a-zA-Z][a-zA-Z0-9\\-_]*$"
...
paths:
...
  /users:
    get:
      security:
        - OAuth2Password: [ ]
        - OAuth2Code: [ "users:read" ]
      parameters:
        - in: query
          name: top
          type: string
          enum: [new, rank, waiting]
          default: new
        - in: query
          name: query
          type: string
          minLength: 1
          maxLength: 50
      tags:
        - users
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              top:
                type: string
              query:
                type: string
              users:
                type: array
                items:
                  $ref: '#/definitions/Friend'
  /users/{name}:
    parameters:
      - $ref: '#/parameters/pathName'
    get:
      security:
        - OAuth2App: []
        - OAuth2Password: [ ]
        - OAuth2Code: [ "users:read" ]
      tags:
        - users
      responses:
        200:
          description: User data
          schema:
            $ref: '#/definitions/Profile'
        404:
          description: User not found
          schema:
            $ref: '#/definitions/Error'
...
  /themes:
    get:
      security:
        - OAuth2Password: [ ]
        - OAuth2Code: [ "themes:read" ]
      parameters:
        - in: query
          name: top
          type: string
          enum: [new, rank]
          default: new
        - in: query
          name: query
          type: string
          minLength: 1
          maxLength: 50
      tags:
        - themes
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              top:
                type: string
              query:
                type: string
              themes:
                type: array
                items:
                  $ref: '#/definitions/Friend'
    post:
      security:
        - OAuth2Password: [ ]
        - OAuth2Code: [ "themes:write" ]
      tags:
        - themes
      consumes:
        - multipart/form-data
        - application/x-www-form-urlencoded
      parameters:
        - in: formData
          name: name
          type: string
          minLength: 1
          maxLength: 20
          required: true
          pattern: "^[a-zA-Z][a-zA-Z0-9\\-_]*$"
        - in: formData
          name: showName
          type: string
          minLength: 1
          maxLength: 20
          required: true
          pattern: "\\s*\\S+.*"
      responses:
        200:
          description: theme data
          schema:
            $ref: '#/definitions/Profile'
        400:
          description: name already exists
          schema:
            $ref: '#/definitions/Error'
        403:
          description: access denied
          schema:
            $ref: '#/definitions/Error'
  /themes/{name}:
    parameters:
      - $ref: '#/parameters/pathName'
    get:
      security:
        - OAuth2App: []
        - OAuth2Password: [ ]
        - OAuth2Code: [ "themes:read" ]
      tags:
        - themes
      responses:
        200:
          description: Theme data
          schema:
            $ref: '#/definitions/Profile'
        404:
          description: Theme not found
          schema:
            $ref: '#/definitions/Error'
    put:
      security:
        - OAuth2Password: [ ]
        - OAuth2Code: [ "themes:write" ]
      tags:
        - themes
      consumes:
        - multipart/form-data
        - application/x-www-form-urlencoded
      parameters:
        - in: formData
          name: showName
          type: string
          minLength: 1
          maxLength: 20
          required: true
          pattern: "\\s*\\S+.*"
        - in: formData
          name: isDaylog
          type: boolean
          default: false
        - in: formData
          name: privacy
          type: string
          enum: [all, invited, registered]
          required: true
        - in: formData
          name: title
          type: string
          maxLength: 500
          default: ''
        - in: formData
          name: showInTops
          type: boolean
          default: false
      responses:
        200:
          description: theme data
          schema:
            $ref: '#/definitions/Profile'
        403:
          description: access denied
          schema:
            $ref: '#/definitions/Error'
...